- Refactored the primary key of each model class (except Meal) to be a long (bigint) instead of int. - Refactored the `Quantity` property in `RecipeIngredient.cs` to be of type double instead of int. - Performed a database migration.

+++ Models/RecipeIngredient.cs
@@ -9,13 +9,13 @@ namespace MEalAPI.Models
     public class RecipeIngredient : BaseEntity
     {
         [Key]
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         [Required]
         public string Units { get; set; } = null!;
 
         [Required]
-        public int Quantity { get; set; }
+        public double Quantity { get; set; }
 
         public bool IsStaple { get; set; }
